refactor(harness): repoint door-eligible shell imports (runtime/ports/spi.prompt_chrome); shell ledger 10→8
