+++ surfaces/interactive_shell/runtime/turn_seams.py
@@ -14,8 +14,7 @@ from typing import Any, Protocol
 from rich.console import Console
 
 from core.agent_harness import OutputSink, ToolCallingTurnResult
-from core.agent_harness.ports import AnswerRequest
-from core.agent_harness.turns.gather_observation import GatheredEvidence
+from core.agent_harness.ports import AnswerRequest, GatheredEvidence
 from core.agent_harness.turns.turn_plan import TurnPlan
 from core.execution import ToolExecutionHooks
 from surfaces.interactive_shell.session import Session
